updated for version 7.4.141 Problem:    Problems when building with Borland: st_mode is signed short;             can't build with Python; temp files not ignored by Mercurial;             building with DEBUG doesn't define _DEBUG. Solution:   Fix the problems. (Ken Takata)

+++ src/os_win32.c
@@ -2960,7 +2960,7 @@ mch_getperm(char_u *name)
     int		n;
 
     n = mch_stat(name, &st);
-    return n == 0 ? (long)st.st_mode : -1L;
+    return n == 0 ? (long)(unsigned short)st.st_mode : -1L;
 }
